# internal/model/model.go
// Package Openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.9.1 DO NOT EDIT.
package model

import (
	"time"
)

const (
	BasicAuthScopes = "BasicAuth.Scopes"
)

// Error defines model for Error.
type Error struct {
	Code    int32  `json:"code"`
	Message string `json:"message"`
}

// FoodPost defines model for FoodPost.
type FoodPost struct {
	Address1       *string    `json:"address1,omitempty"`
	Address2       *string    `json:"address2,omitempty"`
	City           *string    `json:"city,omitempty"`
	Country        *string    `json:"country,omitempty"`
	Date           *time.Time `json:"date,omitempty"`
	Experience     string     `json:"experience"`
	FoodItems      []string   `json:"foodItems"`
	Id             string     `json:"id"`
	Notes          *string    `json:"notes,omitempty"`
	RestaurantName string     `json:"restaurantName"`
	State          *string    `json:"state,omitempty"`
	Zip            *string    `json:"zip,omitempty"`
}

// FoodPostsList defines model for FoodPostsList.
type FoodPostsList []FoodPost

// ListFoodPostsParams defines parameters for ListFoodPosts.
type ListFoodPostsParams struct {
	RestaurantName *string `json:"restaurantName,omitempty"`
}

// CreateFoodPostJSONBody defines parameters for CreateFoodPost.
type CreateFoodPostJSONBody FoodPost

// CreateFoodPostParams defines parameters for CreateFoodPost.
type CreateFoodPostParams struct {
	RestaurantName *string `json:"restaurantName,omitempty"`
}

// CreateFoodPostJSONRequestBody defines body for CreateFoodPost for application/json ContentType.
type CreateFoodPostJSONRequestBody CreateFoodPostJSONBody